CT-1830 ensure we only retrieve accepted responder assignments
We had a bug where in some situations, when retrieving "responder_assignment" on
a case we would return an assignment that hadn't been accepted, which led to
errors when trying to, for example, trying to send an email to the responding
user assigned to a case.

This fix is being abandoned, however, as it seems that this was changed
intentionally as part of ticket CT-931 to allow responders to accept
already-rejected cases. This branch is being left in case we want to revisit
fixing once we better understand how the system should behave (e.g. should
responders be allowed to accept rejected cases that have been assigned to
another team?), and a fix focused on pretenting emails notifications from being
sent to teams that haven't accepted the case is being checked-in instead.